feat: add agent and thread workspace and knowledge file CRUD
Endpoints are added to CRUD files for a general workspace for both
agents and threads. Endpoints are also added to CRUD files for knowledge
in both agents and threads.

When knowledge is added or removed from knowledge, then controller will
kickoff ingestion of the knowledge. An endpoint is added to manually
kickoff ingestion as well.

When deleting an agent or a thread, the corresponding workspaces are
cleaned up.

For now, the `KNOWLEDGE_BIN` environment variable is required for
ingestion.

Signed-off-by: Donnie Adams <donnie@acorn.io>