fix: introduce ingestion and sync custom resources
Using these objects to indicate when ingestion or upload syncing should
occur will dramatically decrease, if not completely remove, the
conflicts seen when uploading a bunch of files at once.

Instead of trying to update statuses in the API layer, a new object is
created and used to determine if and when an ingestion or sync should
occur.

Signed-off-by: Donnie Adams <donnie@acorn.io>