feat: add auth support
This PR adds in support for authentication to the
admin side of Otto. This includes a sign-in page,
user model, me routes, and a AuthContext for rendering
the authenticated user state.

This also includes a change to the API to change
how the default routing was being handled. This
was written by donnie@acorn.io.

Signed-off-by: tylerslaton <mtslaton1@gmail.com>
Co-authored-by: Donnie Adams <donnie@acorn.io>