feat: overhaul sidebar and header
The sidebar now uses the ShadCN component. This
reduces code complexity by a lot. Throughout the
applications the various components for this are
used and overall styling has changed quite a bit.

As such, the old sidebar components are no longer
needed and are deleted.

In addition, the header now has breadcrumbs which
are used to display the current page in a tree
format replacing the previous method of just
displaying a heading.

Finally, the sidebar has been updated to have a
settings menu which contains a link to the OAuth
applications page.

Signed-off-by: tylerslaton <mtslaton1@gmail.com>