fix: use different key for tool reference getters
This is a workaround to make the SWR caching work
nicely with the getToolReferences function. The reason we do this
is because otherwise the SWR cache will not be invalidated between
the two function calls and thus break type safety since types in
TypeScript are erased at runtime. as-map=true does not exist in
the API.

Signed-off-by: tylerslaton <mtslaton1@gmail.com>