feat: add error boundary
This adds an error boundary for the _auth routes.
By adding this, errors are handled in a more consistent
manner in a way that is interactable for the user.

In addition this removes the sonners that would be
shown when API errors occurred. If they are blocking
errors, the user will now be directed to the error
boundary. We'll want things to be more specific
in the future, but this is a great start.

Signed-off-by: tylerslaton <mtslaton1@gmail.com>