fix: reload tool references every hour
If the tool reference doesn't change, but the tool definition does, then
that tool definition will never be pulled into otto8 because the
generation of the tool reference won't change. This change adds a "last
checked" timestamp so that we will check for new tool definitions at
least every hour.

Signed-off-by: Donnie Adams <donnie@acorn.io>