chore: update the user API to be usable
The user API was migrated from the Gateway and wasn't used outside
getting and listing. This change ensures that update and delete work,
especially in regard to changing user roles.

One note: this change does make it possible to change the "nobody" user
when authentication is turned off. However, the role will be switched
back to admin automatically.

Signed-off-by: Donnie Adams <donnie@acorn.io>