enhance: improve controller startup and shutdown
The controller tries to release its lease when it shuts down. However,
the server was shut down at the same time, which made it impossible
for the controller to release its lease. After this change, the startup
and shutdown of the controller is more managed to allow for the lease to
be released.

Signed-off-by: Donnie Adams <donnie@acorn.io>