feat: add support for OAuth PATs (#1387)
feat: add support for OAuth prompt tokens

Instead of configuring and using OAuth, an app can specify that it
supports using tokens. If this is the case, then the tool's credential
should list some number of "prompt_tokens" and, optionally,
"prompt_vars" and the user will be prompted for those. If the oauth2
credential tool should prompt the user for a token instead of using
OAuth, then Obot will not pass the environment variables that feed
the URLs to the tool.

A side effect of this change is that OAuth apps no longer default to
global.

Signed-off-by: Donnie Adams <donnie@acorn.io>
Co-authored-by: Ivy <ivy.jeong@gmail.com>