enhance: stop trigger when tool reference not updated (#1867)
enhance: stop trigger when tool reference not updated

The existing logic would update the tool reference every hour, even when there
were no updates to the tool reference because we would track the last checked
time on the object itself. This would cause a large storm of object processing
because runs, threads, agents, workflows, and workflow executions would all be
triggered in this case.

This change ensures that tool references only trigger related objects when the
tool reference itself changes. In addition to the changes that are explicitly
added here, changes to nah were also required. This change bumps that dependency
as well.

Signed-off-by: Donnie Adams <donnie@acorn.io>