feat: add the ability for a user to remove themselves (#2416)
feat: add the ability for a user to remove themselves

This change adds the backend work to allow a user to delete themselves
and have all of their Obots, threads, runs, credentials, sessions, etc.
cleaned up. This change also ensures that these items are cleaned up
when an admin deletes another user.

The cleanup is accomplished asynchronously, in the controller. This
seemed like a prudent way to do the cleanup for two reasons:
1. To make the API as speedy as possible; and
2. To centralize the cleanup so that other objects could be cleaned up
easily.

There is a slight issue with this: the order in which the items are
cleaned up and the user is removed from the database. There could be
situations where the starting of the cleanup is successful, but the
removal of the user fails. Or visa-versa. This change will start the
cleanup of the objects before trying to remove the user from the
database. This seemed like the most reasonable thing to do because the
deletion of the user could be retried in this scenario. In the reverse
case, if the removal of user object from the database was successful,
but the starting of the cleanup failed, then there is really no way to
start the cleanup again because the user is gone.

Signed-off-by: Donnie Adams <donnie@acorn.io>