enhance: add encryption to PII in users and identities (#2407)
To make this work, some hashed values are also added to the database to
ensure that getting users and identities still works as expected. That
is, we still need to get/update user by username. Therefore, we store a
hash of the username in the database for selecting. Similar for some
other fields.

Another field was added to users and identities: encrypted. This field
tracks if a user or identity is encrypted in the database for migration
and enabling encryption when it was off to begin with. This new field is
really an implementation detail, which means that all
gets/updates/creates for users and identities should be done through the
client forever more.

Necessary migrations are also added for the hashed fields that are now
required for users and identities.

Signed-off-by: Donnie Adams <donnie@acorn.io>