fix: allow user to delete their account from the UI (#2489)
There were a couple issues keeping a user from being able to do this.
First, when encryption is enabled, the user was not able to remove their
account because it could not be found. This change switches the delete
user method to use the hashed_username field.

Secondly, when the user successfully removed their account, the cookie
they used to authenticate would still be valid. The UI would immediately
call /api/me and the user would be authenticated and a new user
created. This change tells the browser to remove the cookie so this does
not happen.

Signed-off-by: Donnie Adams <donnie@acorn.io>