feat: memory (#2586)
* feat: add memory api

Add an Obot API for managing project scoped sets of memories.

Signed-off-by: Nick Hale <4175918+njhale@users.noreply.github.com>

* feat: add memory management dialog to user ui

Add a dialog to the user UI that allows users to view and delete
project memories. The modal is accessible from a button in the sidebar as
well as a button next to `Add Memory` tool calls in the chat component.

Signed-off-by: Nick Hale <4175918+njhale@users.noreply.github.com>

---------

Signed-off-by: Nick Hale <4175918+njhale@users.noreply.github.com>