Add support for creating, managing, and consuming Agent Templates (#2828)
* feat: backend support for agent templates

Make backend changes required to support Agent Templates.

This includes a new public-facing `ProjectTemplate` resource and APIs to
interact with it. Under the hood `ProjectTemplates` are implemented as
an extension of the `Thread` and `ThreadShare` storage types, so makes use
of much of the old "copy Obot" logic still in place.

Signed-off-by: Nick Hale <4175918+njhale@users.noreply.github.com>

* feat: support agent templates in the user ui

Add support for adding, managing, publishing, and consuming Agent Templates to
the user UI.

This change also overhauls the Agent Catalog to be template-backed. It
implements the same search-based UX pattern set by the MCP Server Catalog.

Signed-off-by: Nick Hale <4175918+njhale@users.noreply.github.com>

---------

Signed-off-by: Nick Hale <4175918+njhale@users.noreply.github.com>