chore: make the scope slightly more narrowly scoped (#3033)
It is possible to add the same MCP server configuration twice, but this
won't create a new MCP server instance. This change will ensure that a
separate server instance is created.

This does create a new problem: conflicting tool names. However, that is
easier to fix on the fly because the tools are dynamically generated.

Signed-off-by: Donnie Adams <donnie@acorn.io>