fix: allow authz when public share id not found (#3140)
Allow authz to succeed when the `public_share_id` is not found. This
allows API requests to proceed and return a 404 when users are trying to
access a chatbot from an inactive share link (e.g. the chatbot was
disabled) and causes the UI to load the generic "page not found" page
instead of the "403" page. This change also adds details to the "page
not found" page explaining that the chatbot at the given path does not
exist.

Note: The two routes affected by this change are `GET` and `POST` on `/api/shares/{public_share_id}`
which both respond with a 404 when no share exists with the given `public_share_id`.

Addresses: https://github.com/obot-platform/obot/issues/2792

Signed-off-by: Nick Hale <4175918+njhale@users.noreply.github.com>