fix: address various oauth client validation issues (#3273)
This change will no longer validate grant and response types when
registering an oauth client. It seems that oauth clients don't respect
the supported fields on the oauth-authorization-server endpoint.

Support for oauth_client_basic token authentication type
is added.

A change is made such that if the client uses a token authentication
type of none, then they must use PKCE when doing the authorization_code
flow.

A bug is addressed where the state was not being properly saved.

Finally, we allow all origins in CORS to allow oauth clients running in
the user's browser (like inspector) to work properly.

Signed-off-by: Donnie Adams <donnie@acorn.io>