feat: MCP auditing (#3340)
This change adds auditing through the MCP gateway. As requests come in
from the user's client, the messages and their responses are stored for
querying. Similarly, messages from the server to the user's client as
stored.

Two APIs are added for querying: the audit logs directly and some basic
stats on the MCP server usage. A caller can use query parameters to
control the data that is returned in these APIs.

Signed-off-by: Donnie Adams <donnie@acorn.io>