feat: add support for MCP webhooks and filters (#3473)
Webhooks can be configured at the catalog entry or MCP server level.
The '*' can be used to indicate that a webhook should be called for all
MCP servers or catalog entries. Finally, webhooks can be configured for
some or all MCP methods as well as some or all call identifiers (tool,
prompt, or resource names).

Signed-off-by: Donnie Adams <donnie@acorn.io>