chore: refactor project-based MCP servers (#3509)
chore: refactor project-based MCP servers

This change makes project-based MCP server pointers to a user's MCP
servers. Any changes, including deletes, that happen to the referenced
MCP server happen to this pointer.

Migrations are added so that existing project-based MCP servers are
converted to the new paradigm.

This change also include the UI work necessary to integrate with this new
paradigm in the backend.

Signed-off-by: Donnie Adams <donnie@acorn.io>