enhance: add an MCP server launch API
This API simply ensures that the MCP server is launched, running, and
responding to requests. This should only be used for non-remote MCP
servers because it doesn't handle OAuth. Use the /check-oauth or
/oauth-url APIs for those.

Also, this change moves /check-oauth to a POST request because that's
probably what it should have been the whole time.

Signed-off-by: Donnie Adams <donnie@acorn.io>