fix: use the audit log query to search for users by display name
Because we encrypt personally identifiable information like a user's
display name, we have to list all users from the database and match the
query in code.

This change also improves getting the total number of audit logs for a
search. Instead of issuing a second search, the total count is processed
at the same time as querying for audit logs.

Signed-off-by: Donnie Adams <donnie@acorn.io>