chore: add error message (#3724)
* add error message

* remove inspect statement

* fix: delete project mcp button not working (#3709)

* fix: alleviate block when deleting users when using sqlite (#3734)

When we use sqlite, we have a single database connection. Also, sqlite's
locking is not as sophisticated as others. So, when we're deleting a
user in a transaction that locks the database. Then, we're blocked
when we try to clean up other resources associated with the user from
inside the transaction.

This change alleviates that blocking by cleaning up all user-based
resources in the user-delete controller.

Signed-off-by: Donnie Adams <donnie@acorn.io>

* remove console.log

---------

Signed-off-by: Donnie Adams <donnie@acorn.io>
Co-authored-by: Abdel <abdelhalim.ria@gmail.com>
Co-authored-by: Ivy <ivy@acorn.io>
Co-authored-by: Donnie Adams <donnie@acorn.io>