fix: properly handle requests and responses in audit logs
Similarly to another issue that was fixed previously, the request and
response writer on the message handler were fixed to the first request
and response writer that came through. This change redesigns the message
handling to use a context value for each request. This also allowed for
simplifications in other areas of message handling.

Also included here: calculate the deployment ID as few times as
necessary and pass it as a parameter when possible.

Signed-off-by: Donnie Adams <donnie@acorn.io>