enhance: support MCP servers that require files
Some MCP servers require files in their configuration. This change allows users
to include environment variables in their MCP server configuration and tick the
"file" box. An environment variable that is a file will be replaced with the
file name for the file that is created with the given data.

Signed-off-by: Donnie Adams <donnie@acorn.io>