enhance: make MCP URLs global when possible (#4192)
Instead of each user that connects to an MCP server getting a different URL,
this change introduces the concept of a global URL for an MCP server that any
user can use, so long as they have configured the MCP server.

If a user has configured an MCP server more than once, then the oldest MCP
server will have the "universal" URL, and the remaining servers will have a
unique URL as was before this change. If the user deletes the oldest MCP server,
then the oldest remaining server will have the universal URL. The unique URL
will always work.

If the MCP server has not been configured for a user when they try to connect to
it with this global URL, they will get a 404 error and message explaining that
the server needs to be configured for them first.

Signed-off-by: Donnie Adams <donnie@obot.ai>