fix: detect deleted session when trying to update (#4713)
If a session is deleted from storage, it may still be in the in-memory cache. We
would only detect that it is gone when we try to update its last used time. If
it is determined that the session is gone when we try to update the last used
time, then remove the session from the cache and indicate to the caller that the
session was not found.

Signed-off-by: Donnie Adams <donnie@obot.ai>