fix: block deletion of multi-user servers with dependent composites (#5053)
Return a `409 Conflict` error when a user tries to delete a multi-user
server that is in use by any composite catalog entry or running server.
Include a list of reference to the dependents in the response to make it
easier for the UI to construct actionable dialogs based on the error.

Addresses backend changes for: https://github.com/obot-platform/obot/issues/4967

Signed-off-by: Nick Hale <4175918+njhale@users.noreply.github.com>